fix-canvasid: reset canvas-id instead of retro-fill
- clear canvas-id so it can be pulled from current manifest and saved
  instead of forcing it to be the old url "oculus".
